+++ tests/pmap_test.py
@@ -3016,7 +3016,7 @@ class ShardArgsTest(jtu.JaxTestCase):
     self.assertEqual(len(bufs), 1)
     self.assertEqual(len(bufs[0]), nshards)
     for buf, idx in zip(bufs[0], indices):
-      self.assertAllClose(buf.to_py(), x[idx], check_dtypes=False)
+      self.assertAllClose(np.asarray(buf), x[idx], check_dtypes=False)
 
 
 class ArrayPmapTest(jtu.JaxTestCase):